pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bertytype"
version = "0.1.0"
description = "Local voice dictation powered by VibeVoice and Gemma 4"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "sounddevice>=0.4",
    "numpy>=1.24",
    "pydub>=0.25",
    "pystray>=0.19",
    "Pillow>=10.0",
    "keyboard>=0.13",
    "pywin32>=306",
    "pyperclip>=1.8",
    "pyautogui>=0.9",
    "requests>=2.31",
    "loguru>=0.7",
]

[project.optional-dependencies]
local-stt = [
    "transformers>=4.50",
    "torch>=2.0",
]
dev = ["pytest>=7", "pytest-mock>=3", "mypy>=1.0"]

[project.gui-scripts]
bertytype = "bertytype.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bertytype"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
files = ["src/bertytype"]
